Project Initialization - Configured TailWind; - Added and configured React; - Migrated to TypeScript - Fixed minor bugs
