+++ package.json
@@ -17,6 +17,9 @@
     },
     "dependencies": {
         "@inertiajs/react": "^2.0.3",
-        "react-dom": "^19.0.0"
+        "@types/react": "^19.0.8",
+        "@types/react-dom": "^19.0.3",
+        "react-dom": "^19.0.0",
+        "typescript": "^5.7.3"
     }
 }
