--- resources/js/Pages/Home.jsx
@@ -1,7 +1,0 @@
-export default function Home() {
-    return (
-        <>
-            <h1>Hello user</h1>
-        </>
-    );
-}
