--- resources/js/app.jsx
@@ -1,13 +1,0 @@
-import './bootstrap';
-import { createInertiaApp } from '@inertiajs/react'
-import { createRoot } from 'react-dom/client'
-
-createInertiaApp({
-  resolve: name => {
-    const pages = import.meta.glob('./Pages/**/*.jsx', { eager: true })
-    return pages[`./Pages/${name}.jsx`]
-  },
-  setup({ el, App, props }) {
-    createRoot(el).render(<App {...props} />)
-  },
-})
