+++ resources/ts/Pages/Home.tsx
@@ -1,0 +1,7 @@
+export default function Home() {
+    return (
+        <>
+            <h1 className="bg-amber-400 font-medium text-3xl">Hallo benutzer</h1>
+        </>
+    );
+}
