+++ resources/ts/app.tsx
@@ -1,0 +1,15 @@
+import './bootstrap';
+import { createInertiaApp } from '@inertiajs/react';
+import React from 'react';
+import { createRoot } from 'react-dom/client';
+import "../css/app.css";
+
+createInertiaApp({
+  resolve: name => {
+    const pages = import.meta.glob('./Pages/**/*.tsx', { eager: true })
+    return pages[`./Pages/${name}.tsx`]
+  },
+  setup({ el, App, props }) {
+    createRoot(el).render(<App {...props} />)
+  },
+})
