+++ resources/views/app.blade.php
@@ -1,10 +1,10 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="{{ str_replace('_', '-', app()->getLocale()) }}">
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     @viteReactRefresh
-    @vite('resources/js/app.jsx')
+    @vite('resources/ts/app.tsx')
     @inertiaHead
 </head>
 <body>
